+++ build-tools/processor/src/main/java/co/elastic/apm/compile/processor/generators/instrumentations/InstrumentationBuilderGenerator.java
@@ -1,8 +1,10 @@
-package co.elastic.apm.compile.processor.generators;
+package co.elastic.apm.compile.processor.generators.instrumentations;
 
 import java.util.List;
 
 import javax.lang.model.element.TypeElement;
+
+import co.elastic.apm.compile.processor.generators.Generator;
 
 public class InstrumentationBuilderGenerator implements Generator {
